# app/src/main/java/com/felpeto/rent/adapter/input/controller/handler/dto/Field.java
package com.felpeto.rent.adapter.input.controller.handler.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.ToString;

@Getter
@Builder
@ToString
@EqualsAndHashCode
@NoArgsConstructor
@AllArgsConstructor
public class Field {

  private String name;
  private String message;
}