# app/src/test/resources/openapi-test.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Rent API",
    "description" : "Manage your rental properties",
    "contact" : {
      "name" : "Murilo Felpeto",
      "url" : "https://github.com/murilofelpeto",
      "email" : "murilofelpeto@confidential.com"
    },
    "version" : "1.0.0"
  },
  "paths" : {
    "/v1/properties" : {
      "get" : {
        "summary" : "Get all properties",
        "operationId" : "getProperties",
        "parameters" : [ {
          "name" : "_limit",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "_offset",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "_sort",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return a list of properties",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PropertyResponseDto"
                  }
                }
              }
            }
          },
          "204" : {
            "description" : "empty list"
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create a property",
        "operationId" : "createProperty",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PropertyRequestDto"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Return a created property",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PropertyResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/properties/{id}" : {
      "get" : {
        "summary" : "Get property by it's id",
        "operationId" : "getPropertyById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return a property",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PropertyResponseDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Property not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PropertyResponseDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "Update property by it's id",
        "operationId" : "updateProperty",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PropertyRequestDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Update a property",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PropertyResponseDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ErrorDto" : {
        "type" : "object",
        "properties" : {
          "fields" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Field"
            }
          },
          "message" : {
            "type" : "string"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Field" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "PropertyRequestDto" : {
        "required" : [ "city", "complement", "country", "number", "property_kind", "state", "street_name", "zip_code" ],
        "type" : "object",
        "properties" : {
          "city" : {
            "type" : "string"
          },
          "complement" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "property_kind" : {
            "type" : "string",
            "enum" : [ "single-family", "multi-family", "townhouses", "condominiums", "apartments", "manufactured", "public", "retail", "office", "co-working", "land", "heavy-manufacturing", "light-manufacturing", "warehouses", "distribution-facilities" ]
          },
          "state" : {
            "maxLength" : 2,
            "minLength" : 2,
            "type" : "string",
            "description" : "We expect an abbreviation of the state"
          },
          "street_name" : {
            "type" : "string"
          },
          "zip_code" : {
            "type" : "string"
          }
        }
      },
      "PropertyResponseDto" : {
        "type" : "object",
        "properties" : {
          "city" : {
            "type" : "string"
          },
          "complement" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "propertyKind" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "streetName" : {
            "type" : "string"
          },
          "zipCode" : {
            "type" : "string"
          }
        }
      }
    }
  }
}